+++ tests/testthat/test-build-home-index.R
@@ -32,3 +32,16 @@ test_that("page title defaults to package title", {
     "A test package"
   )
 })
+
+test_that("page title can be overridden", {
+  pkg <- test_path("assets/home-index-rmd")
+  pkg <- as_pkgdown(pkg)
+  pkg$meta <- list(home = list(
+    title = "Such a cool package"))
+
+  expect_equal(
+    as.character(data_home(pkg))[[1]],
+    "Such a cool package"
+  )
+
+})