+++ NEWS.md
@@ -1,5 +1,7 @@
 # pkgdown 1.3.0.9000 (development version)
 
+* One can override the title of the homepage via a `title` field in the home section of config (#957, @maelle).
+
 * The navbar is now automatically hidden with [headroom.js](https://wicky.nillia.ms/headroom.js/).
 
 * The sticky behavior of the navbar is now implemented in pure CSS instead of relying on the 3rd party javascript library (#1016, @bisaloo)